develop > main

+++ src/main/java/com/mate/bbs/vo/GuideTourWriteVO.java
@@ -51,7 +51,7 @@ public class GuideTourWriteVO {
 	 */
 	@Min(value=1 , message="1 이상의 값을 입력해주세요!")
 	@PositiveOrZero(message ="음수를 넣으실 수 없습니다!")
-	private int gdTrPrc;
+	private double gdTrPrc;
 	/**
 	 * 투어 최대 인원수
 	 */
@@ -90,8 +90,6 @@ public class GuideTourWriteVO {
 	 * 투어 작성자의 ID VO
 	 */
 	private String athrId;
-	
-	
 	
 	public String getGdTrPstId() {
 		return gdTrPstId;
@@ -147,10 +145,10 @@ public class GuideTourWriteVO {
 	public void setGdTrMp(String gdTrMp) {
 		this.gdTrMp = gdTrMp;
 	}
-	public int getGdTrPrc() {
+	public double getGdTrPrc() {
 		return gdTrPrc;
 	}
-	public void setGdTrPrc(int gdTrPrc) {
+	public void setGdTrPrc(double gdTrPrc) {
 		this.gdTrPrc = gdTrPrc;
 	}
 	public int getGdTrMxNp() {
@@ -207,5 +205,4 @@ public class GuideTourWriteVO {
 	public void setAthrId(String athrId) {
 		this.athrId = athrId;
 	}
-	
 }
